convert counts to Int32 instead of Int64

+++ src/input.jl
@@ -643,7 +643,7 @@ Returns labeled sparse matrix containing the counts
 function convert_counts(X::AbstractMatrix, features::AbstractVector, barcodes::AbstractVector; unique_features::Bool=true)
     if !(eltype(X) <: Integer)
         @warn "count matrices should be integers, trying to convert from $(eltype(X))"
-        X = convert(AbstractMatrix{Int64}, X)
+        X = convert(AbstractMatrix{Int32}, X)
     end
 
     if unique_features
